@@ -25,7 +25,7 @@ interface LegacyVulnerability {
 
 const isVulnerablePathNew = (monitoredSnapshotPathArray: Array<string>, currentSnapshotPathArray: Array<string> ): boolean => {
 
-    const versionPatternRegex = /@[a-zA-Z0-9-_\.]+$/
+    const versionPatternRegex = /@[a-zA-Z0-9-_.]+$/
 
     if (currentSnapshotPathArray.length === 0){
         return false
