@@ -1,5 +1,5 @@
 import { mockProcessExit } from 'jest-mock-process';
-import * as nock from 'nock';
+import nock from 'nock';
 import * as path from 'path';
 import * as fs from 'fs';
 process.argv.push('-d');
