@@ -1,6 +1,6 @@
 import * as path from 'path';
 import * as fs from 'fs';
-import * as nock from 'nock';
+import nock from 'nock';
 
 import {
   getNewIssues,
